Update .gitignore with precise folder paths
In this update, a slight change was made to the '.idea' folder path in .gitignore file for more accuracy ; also added '.bin/' and '.obj/' folder paths. This change is implemented to further optimize the ignored files and folders, ensuring a cleaner working directory and preventing accidental addition of unnecessary files.